third_party/blink/web_tests/platform/linux-chrome/external/wpt/payment-request/payment-request-disallowed-when-hidden.https-expected.txt

This is a testharness.js-based test.
[FAIL] PaymentRequest.show() cannot be triggered from a hidden context
  assert_equals: expected true but got false
Harness: the test ran to completion.
